FEATURE - Use the EBI workarounds from vf-core
This drops the `_VF1-corrections.scss` and adds `.ebi-vf1-integration` to the `body` element to make use of the component in VF Core.

It does so by adding the class in `wp-content/plugins/vf-ebi-global-footer-container/index.php`, which is perhaps the least bad place as that's where the EBI v1.3 CSS, JS get loaded.